refactor(Servlet): javax.servlet 관련 관련 코드 제거
