+++ tomcat/src/main/java/org/apache/catalina/servlets/DefaultServlet.java
@@ -1,11 +1,5 @@
 package org.apache.catalina.servlets;
 
-import jakarta.servlet.Servlet;
-import jakarta.servlet.ServletConfig;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletRequest;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.http.HttpServletRequest;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
@@ -13,7 +7,8 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.Map;
 import java.util.stream.Collectors;
-import org.apache.catalina.core.HttpResponse;
+import org.apache.catalina.core.request.HttpRequest;
+import org.apache.catalina.core.response.HttpResponse;
 
 /**
  * 정적 리소스를 관리하는 서블릿입니다.
@@ -27,30 +22,17 @@ public class DefaultServlet implements Servlet {
     );
 
     @Override
-    public void init(ServletConfig servletConfig) throws ServletException {
+    public void service(HttpRequest request, HttpResponse response) throws IOException {
 
-    }
-
-    @Override
-    public ServletConfig getServletConfig() {
-        return null;
-    }
-
-    @Override
-    public void service(ServletRequest servletRequest, ServletResponse servletResponse)
-            throws ServletException, IOException {
-
-        HttpServletRequest httpRequest = (HttpServletRequest) servletRequest;
-
-        URL resource = getClass().getClassLoader().getResource("static" + httpRequest.getRequestURI());
+        URL resource = getClass().getClassLoader().getResource("static" + request.getRequestURI());
 
         String fileContent = getFileContent(resource);
         String resourceType = getResourceType(resource);
         String contentType = contentTypes.getOrDefault(resourceType, "text/html;charset=utf-8");
 
-        servletResponse.setContentType(contentType);
-        servletResponse.setContentLength(fileContent.getBytes().length);
-        setResponseBody(servletResponse, fileContent);
+        response.setContentType(contentType);
+        response.setContentLength(fileContent.getBytes().length);
+        response.setResponseBody(fileContent);
     }
 
     private String getFileContent(URL resourceURL) throws IOException {
@@ -68,20 +50,5 @@ public class DefaultServlet implements Servlet {
 
     private String getResourceType(URL resourceURL) {
         return resourceURL.getPath().substring(resourceURL.getPath().lastIndexOf('.') + 1);
-    }
-
-    private void setResponseBody(ServletResponse response, String fileContent) {
-        HttpResponse httpResponse = (HttpResponse) response;
-        httpResponse.setResponseBody(fileContent);
-    }
-
-    @Override
-    public String getServletInfo() {
-        return "";
-    }
-
-    @Override
-    public void destroy() {
-
     }
 }
