+++ tomcat/src/main/java/org/apache/catalina/servlets/LoginServlet.java
@@ -2,18 +2,14 @@ package org.apache.catalina.servlets;
 
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.ServletResponse;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.net.URL;
 import java.util.stream.Collectors;
-import org.apache.catalina.core.HttpResponse;
+import org.apache.catalina.core.request.HttpRequest;
+import org.apache.catalina.core.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,20 +18,17 @@ public class LoginServlet extends HttpServlet {
     private static final Logger log = LoggerFactory.getLogger(LoginServlet.class);
 
     @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        super.doGet(request, response);
-
+    public void doGet(HttpRequest request, HttpResponse response) throws IOException {
         URL resource = getClass().getClassLoader().getResource("static" + request.getRequestURI() + ".html");
         String fileContent = getFileContent(resource);
 
         response.setContentType("text/html");
         response.setContentLength(fileContent.getBytes().length);
-        setResponseBody(response, fileContent);
+        response.setResponseBody(fileContent);
     }
 
     // TODO: 리팩토링 필요 중복 발생
-    private String getFileContent(URL resourceURL) throws IOException {
+    private String getFileContent(URL resourceURL) {
         if (resourceURL != null) {
             File file = new File(resourceURL.getFile());
             try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
@@ -48,34 +41,20 @@ public class LoginServlet extends HttpServlet {
         return "";
     }
 
-    private void setResponseBody(ServletResponse response, String fileContent) {
-        HttpResponse httpResponse = (HttpResponse) response;
-        httpResponse.setResponseBody(fileContent);
-    }
-
     @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        super.doPost(request, response);
+    public void doPost(HttpRequest request, HttpResponse response) {
 
         String account = request.getParameter("account");
         InMemoryUserRepository.findByAccount(account)
-                .ifPresentOrElse(user -> login(request, response, user), () -> redirectTo("/401.html", response));
+                .ifPresentOrElse(user -> login(request, response, user), () -> response.sendRedirect("/401.html"));
     }
 
-    private void login(HttpServletRequest request, HttpServletResponse response, User user) {
+    private void login(HttpRequest request, HttpResponse response, User user) {
         if (user.checkPassword(request.getParameter("password"))) {
             log.info("user : {}", user);
-            redirectTo("/index.html", response);
+            response.sendRedirect("/index.html");
             return;
         }
-        redirectTo("/401.html", response);
-    }
-
-    private void redirectTo(String url, HttpServletResponse response) {
-        try {
-            response.sendRedirect(url);
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
+        response.sendRedirect("/401.html");
     }
 }
